isDeleted 값에 따라 userAPI 변경

+++ back-end/src/services/user-service.ts
@@ -29,9 +29,13 @@ class UserService {
 
     const user = await this.userModel.findByEmail(email);
     if (user) {
-      throw new Error(
-        "이 이메일은 현재 사용중입니다. 다른 이메일을 입력해 주세요."
-      );
+      if (user.isDeleted) {
+        throw new Error("탈퇴한 메일 입니다.");
+      } else {
+        throw new Error(
+          "이 이메일은 현재 사용중입니다. 다른 이메일을 입력해 주세요."
+        );
+      }
     }
 
     const hashedPassword = await bcrypt.hash(password, 10);
@@ -56,6 +60,10 @@ class UserService {
       throw new Error(
         "해당 이메일은 가입 내역이 없습니다. 다시 한 번 확인해 주세요."
       );
+    }
+
+    if (user.isDeleted) {
+      throw new Error("탈퇴한 회원입니다.");
     }
 
     const correctPasswordHash = user.password; // db에 저장되어 있는 암호화된 비밀번호
@@ -92,6 +100,9 @@ class UserService {
     if (!userInfo) {
       throw new Error("해당 ID에 맞는 회원 정보를 불러올 수 없습니다.");
     }
+    if (userInfo.isDeleted) {
+      throw new Error("탈퇴한 회원입니다.");
+    }
     return userInfo;
   }
 
@@ -108,8 +119,14 @@ class UserService {
 
   // 회원 정보 수정
   async setUser(userId: string, toUpdate: UpdateInfo) {
-    if (!(await this.userModel.findById(userId))) {
+    const userInfo = await this.userModel.findById(userId);
+
+    if (!userInfo) {
       throw new Error("가입 내역이 없습니다. 다시 한 번 확인해 주세요.");
+    }
+
+    if (userInfo.isDeleted) {
+      throw new Error("탈퇴한 회원입니다.");
     }
 
     // 업데이트
@@ -127,6 +144,10 @@ class UserService {
       throw new Error("가입 내역이 없습니다. 다시 한 번 확인해 주세요.");
     }
 
+    if (user.isDeleted) {
+      throw new Error("탈퇴한 회원입니다.");
+    }
+
     if (password) {
       const newPasswordHash = await bcrypt.hash(password, 10);
       password = newPasswordHash;
@@ -140,6 +161,10 @@ class UserService {
 
     if (!user) {
       throw new Error("가입 내역이 없습니다. 다시 한 번 확인해 주세요.");
+    }
+
+    if (user.isDeleted) {
+      throw new Error("탈퇴한 회원입니다.");
     }
 
     return await this.userModel.deleteById(userId);
